Added @digitalungdom/bot-identicon
