+++ src/features/Header/index.tsx
@@ -88,7 +88,7 @@ const ConnectedProfileHeaderButton = () => {
 
   return (
     <ProfileHeaderButton
-      avatarSrc={`${Axios.defaults.baseURL}/user/${myProfile?._id}/profile_picture?size=24`}
+      avatarSrc={`${Axios.defaults.baseURL}/user/${myProfile?._id}/profile_picture`}
       firstName={myProfile?.details.firstName || ''}
       lastName={myProfile?.details.lastName || ''}
       loading={myProfile === null}
